+++ .upgrade.yml
@@ -1,6 +1,6 @@
 mappings:
   LDAPAuthenticator: SilverStripe\ActiveDirectory\Authenticators\LDAPAuthenticator
-  LDAPLoginForm: SilverStripe\ActiveDirectory\Authenticators\LDAPLoginForm
+  LDAPLoginForm: SilverStripe\ActiveDirectory\Forms\LDAPLoginForm
   SAMLAuthenticator: SilverStripe\ActiveDirectory\Authenticators\SAMLAuthenticator
   SAMLLoginForm: SilverStripe\ActiveDirectory\Authenticators\SAMLLoginForm
   SAMLSecurityExtension: SilverStripe\ActiveDirectory\Authenticators\SAMLSecurityExtension
