+++ src/Forms/LDAPChangePasswordForm.php
@@ -85,130 +85,4 @@ class LDAPChangePasswordForm extends ChangePasswordForm
             $this->Fields()->unshift($usernameFieldReadonly);
         }
     }
-
-    /**
-     * Change the password
-     *
-     * @param array $data The user submitted data
-     * @return HTTPResponse
-     */
-    public function doChangePassword(array $data)
-    {
-        /**
-         * @var LDAPService $service
-         */
-        $service = Injector::inst()->get(LDAPService::class);
-        $member = Security::getCurrentUser();
-        if ($member) {
-            try {
-                $userData = $service->getUserByGUID($member->GUID);
-            } catch (Exception $e) {
-                Injector::inst()->get('Logger')->error($e->getMessage());
-
-                $this->clearMessage();
-                $this->sessionMessage(
-                    _t(
-                        __CLASS__ . '.NOUSER',
-                        'Your account hasn\'t been setup properly, please contact an administrator.'
-                    ),
-                    'bad'
-                );
-                return $this->controller->redirect($this->controller->Link('changepassword'));
-            }
-            $loginResult = $service->authenticate($userData['samaccountname'], $data['OldPassword']);
-            if (!$loginResult['success']) {
-                $this->clearMessage();
-                $this->sessionMessage(
-                    _t(
-                        'SilverStripe\\Security\\Member.ERRORPASSWORDNOTMATCH',
-                        'Your current password does not match, please try again'
-                    ),
-                    'bad'
-                );
-                // redirect back to the form, instead of using redirectBack() which could send the user elsewhere.
-                return $this->controller->redirect($this->controller->Link('changepassword'));
-            }
-        }
-
-        if (!$member) {
-            if ($this->getSession()->get('AutoLoginHash')) {
-                $member = Member::member_from_autologinhash($this->getSession()->get('AutoLoginHash'));
-            }
-
-            // The user is not logged in and no valid auto login hash is available
-            if (!$member) {
-                $this->getSession()->clear('AutoLoginHash');
-                return $this->controller->redirect($this->controller->Link('login'));
-            }
-        }
-
-        // Check the new password
-        if (empty($data['NewPassword1'])) {
-            $this->clearMessage();
-            $this->sessionMessage(
-                _t(
-                    'SilverStripe\\Security\\Member.EMPTYNEWPASSWORD',
-                    "The new password can't be empty, please try again"
-                ),
-                'bad'
-            );
-
-            // redirect back to the form, instead of using redirectBack() which could send the user elsewhere.
-            return $this->controller->redirect($this->controller->Link('changepassword'));
-        } elseif ($data['NewPassword1'] == $data['NewPassword2']) {
-            // Providing OldPassword to perform password _change_ operation. This will respect the
-            // password history policy. Unfortunately we cannot support password history policy on password _reset_
-            // at the moment, which means it will not be enforced on SilverStripe-driven email password reset.
-            $oldPassword = !empty($data['OldPassword']) ? $data['OldPassword']: null;
-
-            /** @var ValidationResult $validationResult */
-            $validationResult = $service->setPassword($member, $data['NewPassword1'], $oldPassword);
-
-            // try to catch connection and other errors that the ldap service can through
-            if ($validationResult->isValid()) {
-                Security::setCurrentUser($member);
-
-                $this->getSession()->clear('AutoLoginHash');
-
-                // Clear locked out status
-                $member->LockedOutUntil = null;
-                $member->FailedLoginCount = null;
-                $member->write();
-
-                if (!empty($_REQUEST['BackURL'])
-                    // absolute redirection URLs may cause spoofing
-                    && Director::is_site_url($_REQUEST['BackURL'])
-                ) {
-                    $url = Director::absoluteURL($_REQUEST['BackURL']);
-                    return $this->controller->redirect($url);
-                } else {
-                    // Redirect to default location - the login form saying "You are logged in as..."
-                    $redirectURL = HTTP::setGetVar(
-                        'BackURL',
-                        Director::absoluteBaseURL(),
-                        $this->controller->Link('login')
-                    );
-                    return $this->controller->redirect($redirectURL);
-                }
-            } else {
-                $this->clearMessage();
-                $messages = implode('. ', array_column($validationResult->getMessages(), 'message'));
-                $this->sessionMessage($messages, 'bad');
-                // redirect back to the form, instead of using redirectBack() which could send the user elsewhere.
-                return $this->controller->redirect($this->controller->Link('changepassword'));
-            }
-        } else {
-            $this->clearMessage();
-            $this->sessionMessage(
-                _t(
-                    'SilverStripe\\Security\\Member.ERRORNEWPASSWORD',
-                    'You have entered your new password differently, try again'
-                ),
-                'bad'
-            );
-
-            // redirect back to the form, instead of using redirectBack() which could send the user elsewhere.
-            return $this->controller->redirect($this->controller->Link('changepassword'));
-        }
-    }
 }
